Initial sketch of the application
This is a big commit but I wanted to have a functional version of the
site to start with. The basic setup of the Django project and of the
main applicaiton was made.

With this, we set the general layout. The colours were taken without too
much care, and they serve mostly in visually separating the different
sections of the site.

The login link brings us to a login form, but once a user is logged in,
the link does not change to "Logout" like it is intended.

One thing that we'd like to add to the layout would be a CSS trick to
make the footer stick to the bottom and to color the left column
entirely with the current colour.