Fix logout
Currently, the login/logout feature doesn't look like it's working.

Pass the main page's url to the logout view so that users get redirected
there instead of being stuck on a page that looks like the admin
interface.

Make logged out users see a page without the left column, since the list
of machines is irrelevant in this context.

Also, the 'request' object is not automatically passed on to the
templates, so the user name is always empty. Fix this and get the user's
name appropriately.